+++ source/extensions/omni.isaac.core/omni/isaac/core/utils/transformations.py
@@ -14,7 +14,7 @@ from scipy.spatial.transform import Rotation
 from typing import Union, Tuple, Sequence
 
 # omniverse
-from pxr import Gf
+from pxr import Gf, Usd, UsdGeom
 
 # isaacsim
 from omni.isaac.core.utils.rotations import gf_quat_to_np_array
@@ -89,3 +89,89 @@ def tf_matrices_from_poses(
         result[:, :3, 3] = translations
         result[:, 3, 3] = 1
     return result
+
+
+def get_relative_transform(source_prim: Usd.Prim, target_prim: Usd.Prim) -> np.ndarray:
+    """Get the relative transformation matrix from the source prim to the target prim.
+
+    Args:
+        source_prim (Usd.Prim): source prim from which frame to compute the relative transform.
+        target_prim (Usd.Prim): target prim to which frame to compute the relative transform.
+    
+    Returns:
+        np.ndarray: Column-major transformation matrix with shape (4, 4).
+    """
+
+    # Row-major transformation matrix
+    source_to_world_row_major_tf = UsdGeom.Xformable(source_prim).ComputeLocalToWorldTransform(Usd.TimeCode.Default())
+    target_to_world_row_major_tf = UsdGeom.Xformable(target_prim).ComputeLocalToWorldTransform(Usd.TimeCode.Default())
+
+    # Convert to column-major transformation matrix
+    source_to_world_column_major_tf = np.transpose(source_to_world_row_major_tf)
+    target_to_world_column_major_tf = np.transpose(target_to_world_row_major_tf)
+
+    world_to_target_column_major_tf = np.linalg.inv(target_to_world_column_major_tf)
+    source_to_target_column_major_tf = world_to_target_column_major_tf @ source_to_world_column_major_tf
+
+    return source_to_target_column_major_tf
+
+
+def get_translation_from_target(
+    translation_from_source: np.ndarray, source_prim: Usd.Prim, target_prim: Usd.Prim
+) -> np.ndarray:
+    """Get a translation with respect to the target's frame, from a translation in the source's frame.
+
+    Args:
+        translation_from_source (np.ndarray): translation from the frame of the prim at source_path. Shape is (3, ).
+        source_prim (Usd.Prim): prim path of the prim whose frame the original/untransformed translation
+                           (translation_from_source) is defined with respect to.
+        target_prim (Usd.Prim): prim path of the prim whose frame corresponds to the target frame that the returned
+                           translation will be defined with respect to.
+
+    Returns:
+        np.ndarray: translation with respect to the target's frame. Shape is (3, ).
+    """
+
+    translation_from_source_homogenous = np.pad(translation_from_source, ((0, 1)), constant_values=1.0)
+
+    source_to_target = get_relative_transform(source_prim, target_prim)
+
+    translation_from_target_homogenous = translation_from_source_homogenous @ np.transpose(source_to_target)
+    translation_from_target = translation_from_target_homogenous[:-1]
+
+    return translation_from_target
+
+
+def get_world_pose_from_relative(
+    coord_prim: Usd.Prim, relative_translation: np.ndarray, relative_orientation: np.ndarray
+) -> Tuple[np.ndarray, np.ndarray]:
+    """Get a pose defined in the world frame from a pose defined relative to the frame of the coord_prim.
+
+    Args:
+        coord_prim (Usd.Prim): path of the prim whose frame the relative pose is defined with respect to.
+        relative_translation (np.ndarray): translation relative to the frame of the prim at prim_path. Shape is (3, ).
+        relative_orientation (np.ndarray): quaternion orientation relative to the frame of the prim at prim_path.
+                                           Quaternion is scalar-first (w, x, y, z). Shape is (4, ).
+
+    Returns:
+        Tuple[np.ndarray, np.ndarray]: first index is position in the world frame. Shape is (3, ). Second index is
+                                       quaternion orientation in the world frame. Quaternion is scalar-first
+                                       (w, x, y, z). Shape is (4, ).
+    """
+
+    # Row-major transformation matrix from the prim's coordinate system to the world coordinate system
+    prim_transform_matrix = UsdGeom.Xformable(coord_prim).ComputeLocalToWorldTransform(Usd.TimeCode.Default())
+
+    # Convert transformation matrix to column-major
+    prim_to_world = np.transpose(prim_transform_matrix)
+
+    # Column-major transformation matrix from the pose to the frame the pose is defined with respect to
+    relative_pose_to_prim = tf_matrix_from_pose(relative_translation, relative_orientation)
+
+    # Chain the transformations
+    relative_pose_to_world = prim_to_world @ relative_pose_to_prim
+
+    # Translation and quaternion with respect to the world frame of the relatively defined pose
+    world_position, world_orientation = pose_from_tf_matrix(relative_pose_to_world)
+
+    return world_position, world_orientation